+++ .github/workflows/lint-test.yml
@@ -52,7 +52,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install black
+        python -m pip install black==22.12.0
         python -m pip install flake8
         python -m pip install mypy
         python -m pip install pytest
@@ -69,8 +69,8 @@ jobs:
     - name: Type check with Mypy
       run: |
        mypy --version
-       mypy --strict --config-file mypy.ini {logger,host,common,netdev}.py tests/test_{host,common,netdev}.py
+       mypy --strict --config-file mypy.ini ktoolbox/
     - name: Run tests with Pytest
       run: |
        pytest --version
-       pytest tests/test_{host,common,netdev}.py
+       pytest ktoolbox/
