+++ ktoolbox/README.md
@@ -1,0 +1,11 @@
+ktoolbox
+=======
+
+Then `ktoolbox` package contains some basic utilities. It's (current) primary location is
+ocp-traffic-flow-tests ([1]), but it can be reused by just copying the entire directory.
+
+[1] https://github.com/wizhaoredhat/ocp-traffic-flow-tests/tree/main/ktoolbox
+
+If you copy/fork the directory, make sure to not modify the fork. Instead,
+extend the primary location first, and then reimport all files. This is to
+avoid diverging implementations. It requires self-constraint.
