+++ ktoolbox/common.py
