+++ pluginMeasurePower.py
@@ -4,12 +4,13 @@ import time
 from typing import Any
 from typing import Optional
 
-import common
+from ktoolbox import common
+from ktoolbox.logger import logger
+
 import pluginbase
 import task
 import tftbase
 
-from logger import logger
 from task import PluginTask
 from task import TaskOperation
 from testSettings import TestSettings
