+++ tests/test_testConfig.py
@@ -5,7 +5,8 @@ import yaml
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
-import common  # noqa: E402
+from ktoolbox import common  # noqa: E402
+
 import testConfig  # noqa: E402
 
 from tftbase import TestCaseType  # noqa: E402
